extract slot-hashes crate
