+++ sdk/slot-hashes/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "solana-slot-hashes"
+description = "Types and utilities for the Solana SlotHashes sysvar."
+documentation = "https://docs.rs/solana-slot-hashes"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+serde = { workspace = true, optional = true }
+serde_derive = { workspace = true, optional = true }
+solana-hash = { workspace = true, default-features = false }
+
+[dev-dependencies]
+solana-sha256-hasher = { workspace = true }
+
+[features]
+serde = ["dep:serde", "dep:serde_derive"]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
